Does not throw err in the notification job
Fixes #11280, no error return but just log.
